Move the Atreus implementation into a vendor directory
