+++ src/Kaleidoscope-Hardware-Technomancy-Atreus.h
@@ -17,4 +17,4 @@
  */
 
 #pragma once
-#include "kaleidoscope/hardware/Atreus.h"
+#include "kaleidoscope/hardware/technomancy/Atreus.h"
